refactor(threads): useThreadMessaging 主干外移 native 执行目标解析段（2986→2819）
